+++ changelog.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 * Changed: Replaced `msawicki/acf-menu-chooser` with a forked https://github.com/moderntribe/acf-menu-chooser that includes security fixes and is also added to packagist.
 * Updated: ACF (5.12), Tribe Libs (3.4.10), Redirection (5.2.3), Yoast (18.2), TEC (5.14.0.4)
 * Updated: Aligns accordion component with WAI-ARIA standard
+* Fixed: Jetpack sync calls the `all_plugins` filter outside of a screen context, causing fatal errors in our force plugin activation MU plugin when using multisite.
 
 ## 2022.02
 * Updated: Coding standards to v2.1.2, PHP8 sniffer fixes.